requirements.txt
numpy>=1.22.2
pandas>=1.4.2
torch>=2.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
rank-bm25>=0.2.2
scikit-learn>=1.0.2
tqdm>=4.64.0
PyYAML>=6.0
matplotlib>=3.5.1
seaborn>=0.11.2
pytest>=7.1.2
black>=22.3.0
isort>=5.10.1
flake8>=4.0.1
mypy>=0.950 